+++ Backend.csproj
@@ -5,6 +5,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="Bogus" Version="33.0.1" />
         <PackageReference Include="GraphQL" Version="3.3.2" />
         <PackageReference Include="GraphQL.Server.Transports.AspNetCore" Version="4.4.1" />
         <PackageReference Include="GraphQL.Server.Transports.AspNetCore.SystemTextJson" Version="4.4.1" />
@@ -36,6 +37,10 @@
       <_ContentIncludedByDefault Remove="BackendTests\obj\BackendTests.csproj.nuget.dgspec.json" />
       <_ContentIncludedByDefault Remove="BackendTests\obj\project.assets.json" />
       <_ContentIncludedByDefault Remove="BackendTests\obj\project.packagespec.json" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Models" />
     </ItemGroup>
 
 </Project>
