+++ Mocks/MetAPI/Compact.cs
@@ -1,0 +1,7 @@
+namespace Backend.Mocks.MetAPI
+{
+    public class Compact
+    {
+        
+    }
+}
